+++ src/utils/vscodeUtils.ts
@@ -1,4 +1,5 @@
 import * as vscode from 'vscode';
+import { JsonConfiguration } from './types';
 
 const STATUS_BAR_ALIGN = vscode.StatusBarAlignment.Left;
 const STATUS_BAR_PRIORITY = 50;
@@ -14,4 +15,25 @@ export function createStatusBarItem() {
     STATUS_BAR_ALIGN,
     STATUS_BAR_PRIORITY,
   );
+}
+
+export function setContextValue(key: string, value: any): Thenable<void> {
+  return vscode.commands.executeCommand('setContext', key, value);
+}
+
+export function getLaunchConfigIndex(
+  configJson: JsonConfiguration,
+  configName: string,
+) {
+  let configIdx = 0;
+
+  if (configJson) {
+    configJson.configurations.forEach((config) => {
+      if (config.name !== configName) {
+        configIdx++;
+      }
+    });
+  }
+
+  return configIdx;
 }